Move func3d testing into func3d.py
